+++ src/Test/Perf/Runner/CPC-Consumption/ScenarioGenerator.cs
