+++ src/Test/Perf/Runner/CPC-Consumption/TraceManager.cs
